# [GenevaExporter] Update MsgPackTraceExporter for new Enumerate* DS7 APIs