--- a/src/OpenTelemetry.Exporter.Geneva/MsgPackExporter/MsgPackTraceExporter.cs
+++ b/src/OpenTelemetry.Exporter.Geneva/MsgPackExporter/MsgPackTraceExporter.cs
@@ -17,7 +17,6 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading;
 
@@ -241,26 +240,22 @@ internal sealed class MsgPackTraceExporter : MsgPackExporter, IDisposable
             cntFields += 1;
         }
 
-        var links = activity.Links;
-        if (links.Any())
+        cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "links");
+        cursor = MessagePackSerializer.WriteArrayHeader(buffer, cursor, ushort.MaxValue); // Note: always use Array16 for perf consideration
+        var idxLinkPatch = cursor - 2;
+        ushort cntLink = 0;
+        foreach (ref readonly var link in activity.EnumerateLinks())
         {
-            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "links");
-            cursor = MessagePackSerializer.WriteArrayHeader(buffer, cursor, ushort.MaxValue); // Note: always use Array16 for perf consideration
-            var idxLinkPatch = cursor - 2;
-            ushort cntLink = 0;
-            foreach (var link in links)
-            {
-                cursor = MessagePackSerializer.WriteMapHeader(buffer, cursor, 2);
-                cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "toTraceId");
-                cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, link.Context.TraceId.ToHexString());
-                cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "toSpanId");
-                cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, link.Context.SpanId.ToHexString());
-                cntLink += 1;
-            }
-
-            MessagePackSerializer.WriteUInt16(buffer, idxLinkPatch, cntLink);
-            cntFields += 1;
+            cursor = MessagePackSerializer.WriteMapHeader(buffer, cursor, 2);
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "toTraceId");
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, link.Context.TraceId.ToHexString());
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, "toSpanId");
+            cursor = MessagePackSerializer.SerializeAsciiString(buffer, cursor, link.Context.SpanId.ToHexString());
+            cntLink += 1;
         }
+
+        MessagePackSerializer.WriteUInt16(buffer, idxLinkPatch, cntLink);
+        cntFields += 1;
 
         // TODO: The current approach is to iterate twice over TagObjects so that all
         // env_properties can be added the very end. This avoids speculating the size
@@ -274,7 +269,7 @@ internal sealed class MsgPackTraceExporter : MsgPackExporter, IDisposable
         bool isStatusSuccess = true;
         string statusDescription = string.Empty;
 
-        foreach (var entry in activity.TagObjects)
+        foreach (ref readonly var entry in activity.EnumerateTagObjects())
         {
             // TODO: check name collision
             if (CS40_PART_B_MAPPING.TryGetValue(entry.Key, out string replacementKey))
@@ -319,7 +314,7 @@ internal sealed class MsgPackTraceExporter : MsgPackExporter, IDisposable
             cursor = MessagePackSerializer.WriteMapHeader(buffer, cursor, ushort.MaxValue);
             int idxMapSizeEnvPropertiesPatch = cursor - 2;
 
-            foreach (var entry in activity.TagObjects)
+            foreach (ref readonly var entry in activity.EnumerateTagObjects())
             {
                 // TODO: check name collision
                 if (this.m_dedicatedFields.ContainsKey(entry.Key))