## Changes

* Updates `MsgPackTraceExporter` to use the new DS7 "Enumerate" APIs which help with perf.

## Benchmarks

(Modified the benchmark to also use links for this run.)

Before:

|            Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|------------------ |---------:|--------:|--------:|-------:|----------:|
| SerializeActivity | 341.9 ns | 5.34 ns | 4.99 ns | 0.0076 |      96 B |

After:

|            Method |     Mean |   Error |  StdDev | Allocated |
|------------------ |---------:|--------:|--------:|----------:|
| SerializeActivity | 260.0 ns | 4.42 ns | 4.13 ns |         - |

## TODOs

* [X] Appropriate `CHANGELOG.md` updated for non-trivial changes